refactor: add controller to specific folder & docs
