+++ lib/src/modules/data_cache/domain/entities/cache_entity.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/foundation.dart';
 
 import '../dtos/save_cache_dto.dart';
