+++ lib/src/modules/data_cache/presenter/controllers/base_cache_manager_controller.dart
@@ -12,8 +12,8 @@ import '../../domain/usecases/delete_cache_usecase.dart';
 import '../../domain/usecases/get_cache_usecase.dart';
 import '../../domain/usecases/save_cache_usecase.dart';
 
-part 'specifics/prefs_cache_manager_controller.dart';
-part 'specifics/sql_cache_manager_controller.dart';
+part 'specifications/prefs_cache_manager_controller.dart';
+part 'specifications/sql_cache_manager_controller.dart';
 
 class BaseCacheManagerController {
   final GetCacheUsecase _getCacheUsecase;
